data_replayer/requirements.txt
pandas==2.2.0
pyyaml==6.0.1
